+++ xerparser/src/parser.py
@@ -31,17 +31,19 @@ def file_reader(file: str | Path | BinaryIO) -> str:
 
 
 def parser(xer_contents: str) -> dict[str, list]:
-    """Parses the contents of a P6 .xer file and converts it into a Python dictionary object.
+    """
+    Parses the contents of a P6 .xer file and converts it into a
+    Python dictionary object.
 
     Args:
-        file (str): .xer file contents
+        xer_contents (str): .xer file contents
 
     Returns:
-        dict: Dictionary of the xer information and data tables
+        dict: xer information and data tables
     """
     if not isinstance(xer_contents, str):
         raise TypeError(
-            f"TypeError: xer_contents argument must be <class 'str'>; got {type(xer_contents)}"
+            f"TypeError: xer_contents must be <class 'str'>; got {type(xer_contents)}"
         )
 
     if not xer_contents.startswith("ERMHDR"):
@@ -49,9 +51,10 @@ def parser(xer_contents: str) -> dict[str, list]:
 
     table_delimiter = "%T\t"
     tables = xer_contents.split(table_delimiter)
+    xer_data: dict[str, list] = {}
 
     # The first row in the xer file includes file export information
-    xer_data = {"ERMHDR": tables.pop(0).strip().split("\t")[1:]}
+    xer_data["ERMHDR"] = tables.pop(0).strip().split("\t")[1:]
     xer_data.update(
         **{name: rows for table in tables for name, rows in _parse_table(table).items()}
     )